feat: move shared components to shared
