+++ crates/everybody-codes-2025/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod quests;
-pub mod shared;
+
