--- crates/everybody-codes-2025/src/shared/utils.rs
@@ -1,1 +1,0 @@
-
