--- crates/everybody-codes-2025/src/utils.rs
@@ -1,1 +1,0 @@
-
