+++ crates/shared/Cargo.toml
@@ -15,6 +15,8 @@ include.workspace = true
 default = []
 
 [dependencies]
+pastey = "0.2.1"
+pretty_assertions = "1.4.1"
 
 [dev-dependencies]
