+++ crates/shared/src/utils.rs
